#1281: Decode NIP-19 nprofile entities
